split background from assets
