+++ src/game/mod.rs
@@ -1,6 +1,7 @@
 pub use self::canvas::Canvas;
 
 mod assets;
+mod background;
 mod canvas;
 mod controller;
 mod sprites;
